requirements.txt
fastapi
uvicorn[standard]

python-dotenv
openai

requests
beautifulsoup4

pymysql
python-jose[cryptography]
bcrypt
python-multipart
pydantic

rank-bm25
faiss-cpu
rapidfuzz
numpy
pandas
matplotlib
